Demoapp example CQRS shop (#261)
This PR is originally based on the #253 and implements the concept of
the shop that uses Pulsar as message broker in CQRS pattern. To comply
to this pattern we should have strong and precise division into
`Command` and `Query` parts. In the following example the message broker
is used as `EventBus` and `MessageBus` and creation of Commands and
Events made easy thanks to Pulsar's _mltitenancy_.

As each service in our app is connected to the message broker and sends
commands and pushes and queries events from it that means that it would
be reasonable to create a namespace for each service in both `Commands`
and `Events` tenants and to have events and commands as topics inside of
each namespace-service.

There's also a possibility to configure the load of the demoapp by
fine-tuning parameters that will influence on the number subscriptions,
consumers per subscriptions, producers and number sent messages/s for
each topic load category. There, topic load category is just simple
division into 4 different load categories where the topic assigned the
category depending on the context of the topic (I mean I don't thing
that something like `AddCartItem` would have 1 message/s or smth similar
xd. But life isn't fare also)

End structure looks like:

![image](https://github.com/tealtools/pulsar-ui/assets/47789499/eabc5b65-80d0-422d-beff-3b602cf83163)